requirements.txt
djangorestframework
google-generativeai
Pillow
jellyfish